requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
Django==6.0.1
django-filter==25.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.15.0
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
gunicorn==21.2.0
